Split Scene 1 tests in 2 different files
